+++ Planner/js/audio.js
@@ -1,0 +1,5 @@
+let audioTimer = new Audio("./media/rick_and_morty_sun.mp3");
+
+export function audio() {
+    audioTimer.play();
+}
