+++ Planner/js/main.js
@@ -1,5 +1,6 @@
 import { diffDates, diffToHtml } from "./calcdate.js";
 import { formatError } from "./utils.js"
+import { timer } from "./timer.js"
 
 const dateCalcForm = document.getElementById("datecalc");
 const dateCalcResult = document.getElementById("datecalc__result");
@@ -21,30 +22,5 @@ function handleCalcDates(event) {
     }   
 }
 
-let result = document.getElementById('timer__result');
-let startTimer = document.getElementById('startTimer');
-startTimer.addEventListener('click', timer);
-let stopTimer = document.getElementById('stopTimer');
-stopTimer.addEventListener('click', stoptimer)
-let count = 0;
-let timerID;
-  function timer(){
-      count = parseInt(document.getElementById('inputTime').value);
-      if (count != 0 && !isNaN(count)) {
-          timerID = setInterval(() => {
-              if (count == 0) {
-                  clearInterval(timerID)
-              } else {
-                  --count;
-                  result.innerHTML = `Осталось :  ${count} секунд`;
-              }
-        
-          }, 1000);
-      } else {
-        result.innerHTML = formatError("Необходимо ввести число больше нуля");
-      }
-  }
-  
-  function stoptimer(){
-    clearInterval(timerID);
-  }
+
+
